+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2020-05-03
+### Added
+- Added option `skipChildCompilers` to skip applying the plugin to child compilers.
+
+## [2.1.4] - 2020-03-13
+### Fixed
+- Allow for multiple included/excluded chunks
+
 ## [2.1.3] - 2019-10-13
 ### Fixed
 - Ignore package.json files that don't have a `name` field when searching for a package's `package.json` file.